+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.143.0] - Not released
+### Added
+- Hashtag autocomplete in post and comment editors. Start typing `#` followed by
+  letters to see matching hashtags from your posts and other users' posts.
+  Supports non-Latin characters in hashtags.
+
 ## [1.142.0] - 2025-11-30
 ### Added
 - Show found users and groups on the search results page.